requirements.txt
# USA Triathlon Talent ID Pipeline Dependencies
# Pin exact versions for reproducible builds
# Requires Python 3.11+

# Data handling
pandas==2.3.0
sqlalchemy==2.0.34  # upgraded to include Python 3.11 enum bugfix
psycopg2-binary==2.9.10

# Web scraping
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.26.1
webdriver-manager==3.8.5

# Fuzzy matching
rapidfuzz==3.5.2

# Excel/CSV output
openpyxl==3.1.2

# Environment management
python-dotenv==1.0.0

# PDF extraction
pdfplumber==0.10.3

# Development tools
pytest==7.4.3
black==23.11.0
isort==5.12.0
flake8==6.1.0
pre-commit==3.7.0


aiohttp==3.9.4
azure-ai-openai==1.0.0
